contracts-bedrock: remove fund dev accounts (#3285)
* contracts-bedrock: always fund dev accounts

Reduces config complexity

* op-chain-ops: clean up genesis package
