+++ packages/contracts-bedrock/deploy-config/devnetL1.json
@@ -22,7 +22,6 @@
 
   "optimismBaseFeeRecipient": "0xBcd4042DE499D14e55001CcbB24a551F3b954096",
   "optimismL1FeeRecipient": "0x71bE63f3384f5fb98995898A86B02Fb2426c5788",
-  "fundDevAccounts": true,
 
   "deploymentWaitConfirmations": 1
 }
